Fixed issue with paths not working on Windows OS
The path was assumed to have forward slashes when it was being split. This resulted in the path being wrong on Windows where paths are backslashes. The fix was to use the Node.js path library and get the OS specific separator. This fixes the issue on Windows and Unix operating systems which use a forward slash are unaffected.

The error before applying this fix was:

```
node_modules\gulp-slate\index.js:189
throw new PluginError(
FileOpenError: could not open include includes/_errors.md
```